requirements.txt
ultralytics
opencv-python
matplotlib
pandas
pyyaml
tqdm